Update on "fake_quant: make qparams shape consistent"
Summary:

Before this diff, scale+zp were initialized to tensors
with a single dimension and 1 element, and then switched
to scalar tensors after the first forward.

This diff makes the shape stay consistent.  This should fix
an issue reported when saving/loading models, which crashes
on this inconsistent shape.

Test Plan:

```
python test/test_quantization.py TestFakeQuantizePerTensor.test_fake_quant_preserves_qparam_shapes_for_activations
```

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D21605532](https://our.internmc.facebook.com/intern/diff/D21605532)

[ghstack-poisoned]

+++ torch/quantization/observer.py
@@ -189,10 +189,10 @@ class _ObserverBase(ObserverBase):
         # consistent with default values in FakeQuantize.
         if len(scale.shape) == 0:
             # TODO: switch to scale.item() after adding JIT support
-            scale = torch.tensor([float(scale)])
+            scale = torch.tensor([float(scale)], dtype=scale.dtype)
         if len(zero_point.shape) == 0:
             # TODO: switch to zero_point.item() after adding JIT support
-            zero_point = torch.tensor([float(zero_point)])
+            zero_point = torch.tensor([int(zero_point)], dtype=zero_point.dtype)
 
         return scale, zero_point
